MAINT: remove changelog (#66)
We're currently removing all change logs as these haven't been useful or well-maintained yet. We're working on a new solution for presenting this information which will be in place by the time we transition to our beta release. In the meantime, we'll present the most important changes in our release notes, and are available to help users and developers transition between releases on Slack, the QIIME 2 Forum, or repository issue trackers.