BUG: Preserve bar width when tax. level changed
This is done by passing bar width to init(), which in turn is now
done by a "state" object that init() accepts as an argument.

It seems like some of the open problems in the bar plot visualization
(e.g. #91) could be addressed by storing more information about the
visualization's state. (Not sure just passing this around between
functions is the *best* way to structure things, though.)