BUG: Make color scheme persist across tax levels
Previously, changing the taxonomic level would revert the color
scheme back to the default (schemeAccent). Now, it should stick
with whatever the user had selected.

This is done by using the new "state" option of init(). I imagine we
could also use this in the future for outstanding issues like #87.